brakes work current fixed
current changed to float

+++ Steering_Wheel_Board/Constants.h
@@ -55,7 +55,7 @@ bool sdFailure;
 float soc;
 int dcl;
 float cellLowV;
-int currentDraw;
+float currentDraw;
 int cellHighTemp;
 bool currentFail;
 bool voltageFail;
